requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
sqlparse==0.4.4
pytz==2023.3
asgiref==3.7.2
requests==2.31.0
python-dateutil==2.8.2
django-cors-headers==4.3.1
django-environ==0.11.2
